+++ api/tests/test_payload_signature.py
@@ -4,7 +4,7 @@ from django.test import TestCase
 from django.test.client import RequestFactory
 from django.contrib.auth.models import User
 
-from api.host_authentication import SignatureValidation
+from api.authentication import SignatureValidation
 import rest_framework
